Support nested migration paths
`re.match` strictly compares regexes starting at the beginning of the
test string. As a result of this `filter_migration_files` is unable to
location migrations in django applications that are not installed in the
root of the git repository.

Switching to `re.search`, which will look for the pattern anywhere in
the test string, resolves the issue.